+++ src/backend/access/transam/xlogrecovery.c
@@ -812,8 +812,8 @@ InitWalRecovery(ControlFileData *ControlFile, bool *wasShutdown_ptr,
 					(errmsg("entering standby mode")));
 		else if (recoveryTarget == RECOVERY_TARGET_XID)
 			ereport(LOG,
-					(errmsg("starting point-in-time recovery to XID %llu",
-							(unsigned long long) recoveryTargetXid)));
+					(errmsg("starting point-in-time recovery to XID %" PRIu64,
+							recoveryTargetXid)));
 		else if (recoveryTarget == RECOVERY_TARGET_TIME)
 			ereport(LOG,
 					(errmsg("starting point-in-time recovery to %s",
@@ -880,25 +880,25 @@ InitWalRecovery(ControlFileData *ControlFile, bool *wasShutdown_ptr,
 							LSN_FORMAT_ARGS(checkPoint.redo),
 							wasShutdown ? "true" : "false"));
 	ereport(DEBUG1,
-			(errmsg_internal("next transaction ID: %llu; next OID: %u",
-							 (unsigned long long) U64FromFullTransactionId(checkPoint.nextXid),
+			(errmsg_internal("next transaction ID: %" PRIu64 "; next OID: %u",
+							 U64FromFullTransactionId(checkPoint.nextXid),
 							 checkPoint.nextOid)));
 	ereport(DEBUG1,
-			(errmsg_internal("next MultiXactId: %llu; next MultiXactOffset: %llu",
-							 (unsigned long long) checkPoint.nextMulti,
-							 (unsigned long long) checkPoint.nextMultiOffset)));
+			(errmsg_internal("next MultiXactId: %" PRIu64 "; next MultiXactOffset: %" PRIu64,
+							 checkPoint.nextMulti,
+							 checkPoint.nextMultiOffset)));
 	ereport(DEBUG1,
-			(errmsg_internal("oldest unfrozen transaction ID: %llu, in database %u",
-							 (unsigned long long) checkPoint.oldestXid,
+			(errmsg_internal("oldest unfrozen transaction ID: %" PRIu64 ", in database %u",
+							 checkPoint.oldestXid,
 							 checkPoint.oldestXidDB)));
 	ereport(DEBUG1,
-			(errmsg_internal("oldest MultiXactId: %llu, in database %u",
-							 (unsigned long long) checkPoint.oldestMulti,
+			(errmsg_internal("oldest MultiXactId: %" PRIu64 ", in database %u",
+							 checkPoint.oldestMulti,
 							 checkPoint.oldestMultiDB)));
 	ereport(DEBUG1,
-			(errmsg_internal("commit timestamp Xid oldest/newest: %llu/%llu",
-							 (unsigned long long) checkPoint.oldestCommitTsXid,
-							 (unsigned long long) checkPoint.newestCommitTsXid)));
+			(errmsg_internal("commit timestamp Xid oldest/newest: %" PRIu64 "/%" PRIu64,
+							 checkPoint.oldestCommitTsXid,
+							 checkPoint.newestCommitTsXid)));
 	if (!TransactionIdIsNormal(XidFromFullTransactionId(checkPoint.nextXid)))
 		ereport(PANIC,
 				(errmsg("invalid next transaction ID")));
@@ -2714,15 +2714,15 @@ recoveryStopsBefore(XLogReaderState *record)
 		if (isCommit)
 		{
 			ereport(LOG,
-					(errmsg("recovery stopping before commit of transaction %llu, time %s",
-							(unsigned long long) recoveryStopXid,
+					(errmsg("recovery stopping before commit of transaction %" PRIu64 ", time %s",
+							recoveryStopXid,
 							timestamptz_to_str(recoveryStopTime))));
 		}
 		else
 		{
 			ereport(LOG,
-					(errmsg("recovery stopping before abort of transaction %llu, time %s",
-							(unsigned long long) recoveryStopXid,
+					(errmsg("recovery stopping before abort of transaction %" PRIu64 ", time %s",
+							recoveryStopXid,
 							timestamptz_to_str(recoveryStopTime))));
 		}
 	}
@@ -2859,16 +2859,16 @@ recoveryStopsAfter(XLogReaderState *record)
 				xact_info == XLOG_XACT_COMMIT_PREPARED)
 			{
 				ereport(LOG,
-						(errmsg("recovery stopping after commit of transaction %llu, time %s",
-								(unsigned long long) recoveryStopXid,
+						(errmsg("recovery stopping after commit of transaction %" PRIu64 ", time %s",
+								recoveryStopXid,
 								timestamptz_to_str(recoveryStopTime))));
 			}
 			else if (xact_info == XLOG_XACT_ABORT ||
 					 xact_info == XLOG_XACT_ABORT_PREPARED)
 			{
 				ereport(LOG,
-						(errmsg("recovery stopping after abort of transaction %llu, time %s",
-								(unsigned long long) recoveryStopXid,
+						(errmsg("recovery stopping after abort of transaction %" PRIu64 ", time %s",
+								recoveryStopXid,
 								timestamptz_to_str(recoveryStopTime))));
 			}
 			return true;
@@ -2903,9 +2903,9 @@ getRecoveryStopReason(void)
 
 	if (recoveryTarget == RECOVERY_TARGET_XID)
 		snprintf(reason, sizeof(reason),
-				 "%s transaction %llu",
+				 "%s transaction %" PRIu64,
 				 recoveryStopAfter ? "after" : "before",
-				 (unsigned long long) recoveryStopXid);
+				 recoveryStopXid);
 	else if (recoveryTarget == RECOVERY_TARGET_TIME)
 		snprintf(reason, sizeof(reason),
 				 "%s %s\n",
